Update jquery.typing to 0.3.2
